+++ rui-runtime/src/commonTest/kotlin/hu/simplexion/rui/runtime/test/manual/BranchTest.kt
@@ -69,9 +69,9 @@ class Branch(
 
     var v0: Int = 1
 
-    var ruiDirty0 = 0
+    var ruiDirty0 = 0L
 
-    fun ruiInvalidate0(mask: Int) {
+    fun ruiInvalidate0(mask: Long) {
         ruiDirty0 = ruiDirty0 or mask
     }
 
@@ -79,7 +79,7 @@ class Branch(
         if (scopeMask and 1 != 0L) return 0L
 
         it as RuiT1
-        if (ruiDirty0 and 1 != 0) {
+        if (ruiDirty0 and 1L != 0L) {
             it.p0 = v0 + 10
             ruiInvalidate0(1)
         }
@@ -91,7 +91,7 @@ class Branch(
         if (scopeMask and 1 != 0L) return 0L
 
         it as RuiT1
-        if (ruiDirty0 and 1 != 0) {
+        if (ruiDirty0 and 1L != 0L) {
             it.p0 = v0 + 20
             ruiInvalidate0(1)
         }
@@ -102,7 +102,7 @@ class Branch(
     override fun ruiPatch(scopeMask: Long) {
         val extendedScopeMask = ruiFragment.ruiExternalPatch(ruiFragment, scopeMask)
         if (extendedScopeMask != 0L) ruiFragment.ruiPatch(extendedScopeMask)
-        ruiDirty0 = 0
+        ruiDirty0 = 0L
     }
 
     fun ruiBranch0(): RuiFragment<TestNode> = RuiT1(ruiAdapter, this, ::ruiEp0, v0 + 10)
