+++ rui-runtime/src/jsMain/kotlin/hu/simplexion/rui/runtime/dom/html/Button.kt
@@ -27,13 +27,13 @@ class RuiButton(
 
     override val receiver = document.createElement("button") as HTMLButtonElement
 
-    var ruiDirty0 = 0
+    var ruiDirty0 = 0L
 
     override val ruiScope: RuiFragment<Node>
         get() = TODO("Not yet implemented")
 
     @RuiPublicApi
-    fun ruiInvalidate0(mask: Int) {
+    fun ruiInvalidate0(mask: Long) {
         ruiDirty0 = ruiDirty0 or mask
     }
 
@@ -43,10 +43,10 @@ class RuiButton(
     }
 
     override fun ruiPatch(scopeMask: Long) {
-        if (ruiDirty0 and 1 != 0) {
+        if (ruiDirty0 and 1L != 0L) {
             receiver.innerText = label
         }
-        if (ruiDirty0 and 2 != 0) {
+        if (ruiDirty0 and 2L != 0L) {
             receiver.onclick = onClick
         }
     }
